Add Python 3 back to the cram tests
The cram tests fail, but actually pip-review appears to work just fine
under Python 3. The tests are only failing because cElementTree does
not install under Python 3.

We will have to figure out why cElementTree is part of the tests, and
whether there is an alternative that can be installed under Python 3.