Rename some variables and remove a comment
argparse.ArgumentParser.parse_known_args splits the arguments in two
lists, the known arguments and the remaining arguments. This is how
the function is described in the official documentation. The original
code that introduced parse_known_args, however, consistently named the
respective lists `args` and `unknown`. The latter is rather confusing
to find as a variable name in code, hence the name changes.

In addition to this, the comment that I removed was a bit misleading,
as it claimed that parse_known_args is an 'internal' method. The
contrary is true: it is a public, documented method in the standard
library which was specifically added for the purpose of argument
forwarding. The code now describes itself well, so the comment is not
needed anymore.