feat: implement --continue-on-fail argument
I decided to go with `--continue-on-fail` instead of `--ignore-failed` because
`-i` and `-I` shortcuts are already in use by pip and also, I don't think
ignoring failures is a good idea. Maybe in the future the implementation
will change so that the returned status code of pip-review will indicate
a failure along with printing a summary of packages that have failed.
That said, I'm OK with changing it `--ignore-failed`.

closes #88